requirements.txt
bidict==0.23.1
certifi==2025.4.26
charset-normalizer==3.4.2
greenlet==3.2.2
h11==0.16.0
idna==3.10
proxmoxer==2.2.0
python-dotenv==1.1.0
python-engineio==4.12.1
python-socketio==5.13.0
requests==2.32.3
simple-websocket==1.1.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
